requirements.txt
drawsvg[all]==2.3.0
libigl==2.4.1
pythreejs==2.4.1
meshplot @ https://github.com/skoch9/meshplot/archive/0.4.0.tar.gz
CairoSVG==2.7.1
pygeodesic==0.1.8
networkx==3.2.1
matplotlib
shapely